% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoDummy.R
\name{autoDummy}
\alias{autoDummy}
\title{Automatically replace factors with dummy features.}
\usage{
autoDummy(task, wl, ...)
}
\arguments{
\item{task}{(\code{Task}) An mlr Task.}

\item{wl}{(\code{WrappedLearner}) An ezmlr WrappedLearner object.}

\item{...}{Additional arguments to pass to \code{\link[mlr]{makeDummyFeaturesWrapper}}}
}
\description{
Checks a task to see if it contains factors or ordered factors, then checks a
learner to see if it is unable to handle factors. If both are true, returns
the learner wrapped with \code{\link[mlr]{makeDummyFeaturesWrapper}} which
perfors one-hot encoding, i.e. creates a separate column for each factor
level with a 1 indicating that observation contained the factor level in
question, and a zero indicating otherwise.
}
