% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy_autoxgb.R, R/Strategy_autoxgb_tuned.R
\name{autoxgb}
\alias{autoxgb}
\alias{autoxgb}
\title{autoxgb strategy}
\description{
The \code{autoxgb} strategy creates an XGBoost learner to fit a specific task. It performs
imputation if any variables are missing, and one-hot encoding if any factor variables
are present. By default, it uses the following hyperparameters:
\itemize{
\item{\code{nrounds = 200}}
\item{\code{eta = 2^-6}}
\item{\code{max_depth = 3}}
\item{\code{colsample_bytree = 0.6}}
\item{\code{subsample = 0.6}}
}

Available parameters for use in in \code{strategy.pars}:
\itemize{
\item{\code{par.vals} (\code{list})} {List of hyperparameter values to override defaults.}
\item{\code{impute.pars} (\code{list})} {List of parameters to pass to \code{makeImputeWrapper}}
\item{\code{dummy.pars} (\code{list})} {List of parameters to pass to \code{makeDummyFeaturesWrapper}}
}

The \code{autoxgb_tuned} strategy creates an XGBoost learner to fit a specific
task and tunes it for improved performance.

Available parameters for use in in \code{strategy.pars}:
\itemize{
\item{\code{impute.pars} (\code{list})} {List of parameters to pass to \code{makeImputeWrapper}}
\item{\code{dummy.pars} (\code{list})} {List of parameters to pass to \code{makeDummyFeaturesWrapper}}
\item{\code{tune.pars} (\code{list})} {list of parameters to pass to \code{makeTuneWrapper}}
}
}
