% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diagnose}
\alias{diagnose}
\title{Diagnose a learner, strategy, or model with respect to a task.}
\usage{
diagnose(obj, task, plots = NULL, threshVsPerf = NULL, ROCCurve = NULL,
  residuals = NULL, learningCurve = NULL, filterValues = NULL,
  hyperParsEffect = NULL, optPath = NULL, tuneMultiCritResult = NULL,
  PartialDependence = NULL, calibration = NULL, ...)
}
\arguments{
\item{obj}{(\code{learner}, \code{strategy}, or \code{model})\cr
The object to diagnose}

\item{task}{(\code{task})\cr
If \code{obj} is a \code{learner} or a \code{strategy}, a \code{task} must be specified on which
to evaluate performance.}

\item{plots}{(\code{list})\cr
List of plots to produce. If not specified, the function will choose a set
of appropriate plots based on the object and task. Will be overridden if
any of the individual plot parameters is set to \code{TRUE} or \code{FALSE}.}

\item{threshVsPerf}{(\code{logical})\cr
Plot threshold vs. performance curve using \code{plotThreshVsPerf}? Only
available for classification tasks.}

\item{ROCCurve}{(\code{logical})\cr
Plot receiver operating characteristic (ROC) curve using \code{plotROCCurves}?
Only available for classification tasks.}

\item{residuals}{(\code{logical})\cr
Plot residuals of predictions using \code{plotResiduals}?}

\item{learningCurve}{(\code{logical})\cr
Plot learning curve showing learner performance vs. proportion of data
used, using \code{plotLearningCurve}?}

\item{filterValues}{(\code{logical})\cr
Plot feature importance for a given filter method using \code{plotFilterValues}?}

\item{hyperParsEffect}{(\code{logical})\cr
Plot hyperparameter effect on model performance from tuning data using
\code{plotHyperParsEffect}?}

\item{optPath}{(\code{logical})\cr
Plot optimization path from tuning data using \code{plotOptPath}?}

\item{tuneMultiCritResult}{(\code{logical})\cr
Plot pareto front for results of tuning to multiple performance measures
using \code{plotTuneMultiCritResult}?}

\item{calibration}{(\code{logical})\cr
Plot calibration of probability predictions vs. true incidence using
\code{plotCalibration}?}

\item{...}{Additional arguments to pass to each plot.}

\item{partialDependence}{(\code{logical})\cr
Plot partial dependence of model prediction over each data feature using
\code{plotPartialDependence}?}
}
\description{
Given a \code{task} and either a \code{strategy} or \code{learner}, or else given a trained \code{model},
create diagnostic plots to provide useful information about model performance and
the inner workings of the model.
}
