% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R
\name{makeStrategy}
\alias{makeStrategy}
\title{Specify your own learning strategy.}
\usage{
makeStrategy(id, name = id, task.type = NULL, properties = NULL,
  base.learner = NULL, wrappers = NULL, fun)
}
\arguments{
\item{id}{(\code{character(1)})\cr
Name of the strategy.}

\item{name}{(\code{character(1)})\cr
Long name of the strategy. Default is \code{id}.}

\item{task.type}{(\link{character})\cr
Supported task types of the strategy, e.g. 'classif', 'regr', etc.}

\item{properties}{(\link{character})\cr
Supported learner properties of the strategy, e.g. 'numerics', 'factors', 'missings', etc.}

\item{base.learner}{(\link{character})\cr
Base learner or learners that strategy uses}

\item{wrappers}{(\link{character})\cr
Wrapper or wrappers that strategy uses}

\item{fun}{(\code{function(task, pars)})\cr
Takes a \code{Task} and returns a wrapped \code{Learner} customized to the task.
\code{pars} may be added optionally to override default strategy settings.
\describe{
\item{\code{task} (\link{Task})}{The task around which to build the learning strategy.}
\item{\code{pars} (\code{list})}{An optional list of parameters to override the defaults.}
}}
}
\value{
(\link{Strategy}).
}
\description{
Create a strategy which uses data about the task to automatically configure a
wrapped \code{Learner} to train the learner.
}
\examples{
# TBD
}
\seealso{
\link{strategies}
}
