% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWrappedLearner.R
\name{makeWrappedLearner}
\alias{makeWrappedLearner}
\title{Build a wrapped learner using a specified learning strategy.}
\usage{
makeWrappedLearner(strategy, id = strategy, task, strategy.pars = list(),
  learner.pars = list(), ..., predict.type = NULL,
  predict.threshold = NULL, fix.factors.prediction = TRUE, config = NULL)
}
\arguments{
\item{strategy}{(\code{character(1)})\cr
Name (short name) of the strategy to use for training.}

\item{id}{(\code{character(1)})\cr
Learner ID. Default is \code{strategy}.}

\item{task}{(\code{\link[mlr]{Task}})\cr
Task to which the strategy should be applied.}

\item{strategy.pars}{(\code{list})\cr
Optional parameters to pass to the strategy.
These will vary strategy by strategy.}

\item{learner.pars}{(\code{list})\cr
Optional parameters to pass to the wrapped
learner. These are the standard parameters used by each learner in \code{mlr}.}
}
\description{
Follows a specified learning strategy (\code{Strategy}) to create a wrapped learner
customized to a specific task. May include things such as:
\itemize{
\item Preprocessing
\item Imputation of missing values
\item Feature selection
\item Feature filtering
\item Upsampling/downsampling
\item Hyperparameter tuning
}
}
