% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R
\name{Strategy}
\alias{Strategy}
\title{Strategy object.}
\description{
A learning \code{Strategy} builds a wrapped \code{Learner} for training a \code{Task}, which may be
either manually configured or (default) automatically customized to the
characteristics of that \code{Task}.
To see all currently implemented strategies, look at \link{strategies}.

Object members:
\describe{
\item{id (\code{character(1)})}{Name of the strategy.}
\item{name (\code{character(1)})}{Long name of the strategy.}
\item{task.type (\code{character})}{Supported task types of the strategy.}
\item{properties (\code{character})}{Supported learner properties of the strategy.}
\item{fun (\code{function(task, pars)})}{Function that executes the strategy and returns a wrapped \code{Learner}.}
}
}
\seealso{
\link{makeStrategy}
}
