% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train a strategy.}
\usage{
train(strategy, task, max.cores = NULL, ...)
}
\description{
Given a \code{Strategy} and a \code{task}, creates an appropriate \code{Wrapped Learner} and then
trains a model on that wrapped learner. If passed a base mlr learner, will simply
train that learner directly.
}
