% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeWrappedLearner.R
\name{makeWrappedLearner}
\alias{makeWrappedLearner}
\title{Build a wrapped learner using a specified learning strategy.}
\usage{
makeWrappedLearner(strategy, id = strategy, task, predict.type = "response",
  predict.threshold = NULL, fix.factors.prediction = FALSE, ...,
  strategy.pars = list(), learner.pars = list(), config = list())
}
\arguments{
\item{strategy}{(\code{character(1)})\cr
Name of the strategy to use for tuning.}

\item{name}{(\code{character(1)})\cr
Long name of the strategy. Default is \code{id}.}

\item{properties}{(\link{character})\cr
Set of strategy properties.
\describe{
\item{prop1}{Property1}
\item{prop2}{Property2}
}}

\item{fun}{(\code{function(task, pars)})\cr
Takes a \code{Task} and returns a wrapped \code{Learner} customized to the task.
\code{pars} may be added optionally to override default strategy settings.
\describe{
\item{\code{task} (\link{Task})}{The task around which to build the learning strategy.}
\item{\code{pars} (\code{list})}{An optional list of parameters to override the defaults.}
}}
}
\value{
(\link{Strategy}).
}
\description{
Follows a specified learning strategy (\code{Strategy}) to create a wrapped learner
customized to a specific task. May include things such as:
\itemize{
\item Preprocessing
\item Imputation of missing values
\item Feature selection
\item Feature filtering
\item Upsampling/downsampling
\item Hyperparameter tuning
}
}
\examples{
# TBD
}
\seealso{
\link{strategies}
}
